Tests/LibWeb/Text/expected/css/imported-selector-insights-invalidated.txt

before: rgb(0, 128, 0)
cached before import: false
after import: rgb(0, 128, 0)
cached after import: true
after: rgb(255, 0, 0)
